+++ package.json
@@ -16,6 +16,7 @@
     "react-icons": "^4.8.0",
     "react-router-dom": "^6.11.1",
     "react-scripts": "5.0.1",
+    "react-select": "^5.7.3",
     "web-vitals": "^2.1.4"
   },
   "overrides": {
