+++ src/index.css
@@ -67,14 +67,47 @@ body {
 }
 
 
-
-
-
-
 code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
+
+/* Style for Standard Select - Non MDB Select Drop Down List */
+/* USE FOR MEDIA QUERY - SMALLER */
+.birthdate-MDB-input{
+  border: 1px solid rgb(182, 181, 181);
+  height: 4.78vh;
+  width: 10.95vw;
+  border-radius: 1rem;
+  color:rgb(114, 113, 113) ;
+  padding-left: .4vw;
+}
+
+select, select option{
+  background-color: white;
+}
+select.gender-select{
+  border: 1px solid rgb(182, 181, 181);
+  height: 4.78vh;
+  width: 11.5vw;
+  border-radius: 3px;
+  color:rgb(114, 113, 113) ;
+  padding-left: .4vw;
+  margin-left: -3vh;
+  background-color:white ;
+  background:white
+}
+
+select.orientation-select{
+  border: 1px solid rgb(182, 181, 181);
+  height: 4.78vh;
+  width: 13.5vw;
+  border-radius: 3px;
+  color:rgb(114, 113, 113) ;
+  padding-left: .4vw;
+}
+
+
 
 /* ******* Index Page - contains user profile cards and preferences list ******** */
 .Index{
@@ -85,9 +118,13 @@ code {
 /* ***************** User Profile Card ************** */
 
 #profileImage {
-  height: 130px;
-  width: 130px;
-  margin-top: 3vh;
+  /* height: 130px;
+  width: 130px; */
+  margin-top: 1vh;
+  width: 70%;
+  height: 70%;
+  object-fit: cover;
+  object-position: center;
 }
 
 .users {
@@ -194,4 +231,36 @@ code {
   /* width: 12em; */
 }
 
+/* MEDIA QUERIES - to be completed*/
 
+/* USE FOR MEDIA QUERY - SMALLER */
+
+/* 
+.birthdate-MDB-input{
+  border: 1px solid rgb(182, 181, 181);
+  height: 4.78vh;
+  width: 13.5vw;
+  border-radius: 3px;
+  color:rgb(114, 113, 113) ;
+  padding-left: .4vw;
+}
+
+
+.gender-select{
+  border: 1px solid rgb(182, 181, 181);
+  height: 4.78vh;
+  width: 13.5vw;
+  border-radius: 3px;
+  color:rgb(114, 113, 113) ;
+  padding-left: .4vw;
+}
+
+.orientation-select{
+  border: 1px solid rgb(182, 181, 181);
+  height: 4.78vh;
+  width: 13.5vw;
+  border-radius: 3px;
+  color:rgb(114, 113, 113) ;
+  padding-left: .4vw;
+}
+ */
