+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
 import ReactDOM from 'react-dom/client';
-// import 'mdb-react-ui-kit/dist/css/mdb.min.css';
-// import '@fortawesome/fontawesome-free/css/all.min.css';
+import '@fortawesome/fontawesome-free/css/all.min.css';
+import 'mdb-react-ui-kit/dist/css/mdb.min.css';
 import './index.css';
 import App from './App';
 import AuthProvider from './Firebase/context';
